requirements.txt
deta==1.1.0
plotly==5.9.0
python-dotenv==0.21.1
streamlit==1.18.1
streamlit_option_menu==0.3.2
